requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
cssselect==1.1.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.1
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.8
lxml==4.4.2
pkg-resources==0.0.0
pyquery==1.4.1
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
static3==0.7.0
urllib3==1.25.7
